+++ dephell/cache.py
@@ -7,7 +7,7 @@ from typing import List
 
 # app
 from .config import config
-from .utils import cached_property
+from .cached_property import cached_property
 
 
 class BaseCache: