+++ dephell/config/scheme.py
@@ -20,18 +20,18 @@ _TARGET = dict(
 
 # all fields with default value (defaults.py) marked as required.
 SCHEME = {
-    'from': dict(required=False, **_TARGET),
-    'to': dict(required=False, **_TARGET),
-    'and': dict(type='list', schema=_TARGET, required=False, empty=True),
-    'sdist': dict(
+    'from':     dict(required=False, **_TARGET),
+    'to':       dict(required=False, **_TARGET),
+    'and':      dict(type='list', schema=_TARGET, required=False, empty=True),
+    'sdist':    dict(
         type='dict',
         required=True,
         schema={'ratio': dict(type='float', required=True)},
     ),
-    'envs': dict(type='list', required=False, empty=False),
-    'tests': dict(type='list', required=True),
+    'envs':         dict(type='list', required=False, empty=False),
+    'tests':        dict(type='list', required=True),
 
-    'warehouse':    dict(type='string', required=True),
+    'warehouse':    dict(type='list', required=False, empty=False),
     'bitbucket':    dict(type='string', required=True),
     'repo':         dict(type='string', required=False, allowed=REPOSITORIES),
 
@@ -45,13 +45,13 @@ SCHEME = {
     'nocolors':     dict(type='boolean', required=True),
     'filter':       dict(type='string', required=False),
     'traceback':    dict(type='boolean', required=True),
-    'pdb':    dict(type='boolean', required=True),
+    'pdb':          dict(type='boolean', required=True),
 
     # venv
     'venv':     dict(type='string', required=True),
     'dotenv':   dict(type='string', required=True),
     'python':   dict(type='string', required=False),
-    'vars': dict(
+    'vars':     dict(
         type='dict',
         keyschema={'type': 'string'},
         valueschema={'type': 'string'},
@@ -59,14 +59,14 @@ SCHEME = {
     ),
 
     # other
-    'owner': dict(type='string', required=False),
-    'tag': dict(type='boolean', required=False),
-    'cache': dict(
+    'owner':    dict(type='string', required=False),
+    'tag':      dict(type='boolean', required=False),
+    'cache':    dict(
         type='dict',
         required=True,
         schema={
             'path': dict(type='string', required=True),
-            'ttl': dict(type='integer', required=True),
+            'ttl':  dict(type='integer', required=True),
         },
     ),
     'project':      dict(type='string', required=True),