+++ dephell/converters/pip.py
@@ -1,4 +1,5 @@
 # built-in
+from itertools import chain
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Optional
@@ -12,9 +13,9 @@ from pip._internal.req import parse_requirements
 
 # app
 from ..config import config
-from ..controllers import DependencyMaker
+from ..controllers import DependencyMaker, RepositoriesRegistry
 from ..models import RootDependency
-from ..repositories import WareHouseRepo
+from ..repositories import WarehouseLocalRepo
 from .base import BaseConverter
 
 
@@ -44,13 +45,15 @@ class PIPConverter(BaseConverter):
         deps = []
         root = RootDependency()
 
-        warehouse_url = urlparse(config['warehouse']).hostname
-        if warehouse_url in ('pypi.org', 'pypi.python.org'):
-            warehouse_url += '/simple'
+        warehouse_urls = []
+        for url in config['warehouse']:
+            host = urlparse(url).hostname
+            if host in ('pypi.org', 'pypi.python.org', 'test.pypi.org'):
+                warehouse_urls.append('https://{}/simple'.format(host))
 
         finder = PackageFinder(
             find_links=[],
-            index_urls=[warehouse_url],
+            index_urls=warehouse_urls,
             session=PipSession(),
         )
         # https://github.com/pypa/pip/blob/master/src/pip/_internal/req/constructors.py
@@ -77,13 +80,15 @@ class PIPConverter(BaseConverter):
             ))
 
         # update repository
-        if finder.index_urls:
-            finded_host = urlparse(finder.index_urls[0]).hostname
-            if finded_host != urlparse(warehouse_url).hostname:
-                repo = WareHouseRepo(url=finder.index_urls[0])
-                for dep in deps:
-                    if isinstance(dep.repo, WareHouseRepo):
-                        dep.repo = repo
+        if finder.index_urls or finder.find_links:
+            repo = RepositoriesRegistry()
+            for url in chain(finder.index_urls, finder.find_links):
+                repo.add_repo(url=url)
+            for url in config['warehouse']:
+                repo.add_repo(url=url)
+            for dep in deps:
+                if isinstance(dep.repo, RepositoriesRegistry):
+                    dep.repo = repo
 
         root.attach_dependencies(deps)
         return root
@@ -93,23 +98,27 @@ class PIPConverter(BaseConverter):
         lines = []
 
         # get repos urls
-        urls = dict()
+        urls = []
+        paths = []
+        names = set()
         for req in reqs:
-            if isinstance(req.dep.repo, WareHouseRepo):
-                urls[req.dep.repo.name] = req.dep.repo.pretty_url
-
+            if not isinstance(req.dep.repo, RepositoriesRegistry):
+                continue
+            for repo in req.dep.repo.repos:
+                if repo.name in names:
+                    continue
+                names.add(repo.name)
+                if isinstance(repo, WarehouseLocalRepo):
+                    paths.append(repo.path)
+                else:
+                    urls.append(repo.pretty_url)
         # dump repos urls
-        # pip._internal.build_env
-        if len(urls) == 1:
-            _name, url = urls.popitem()
-        elif 'pypi' in urls:
-            url = urls.pop('pypi')
-        else:
-            url = None
-        if url:
-            lines.append('-i ' + url)
-        for url in urls.values():
+        if urls:
+            lines.append('-i ' + urls[0])
+        for url in urls[1:]:
             lines.append('--extra-index-url ' + url)
+        for path in paths:
+            lines.append('--find-links ' + path)
 
         # disable hashes when dir-based deps are presented
         # https://github.com/dephell/dephell/issues/41