+++ dephell/repositories/__init__.py
@@ -1,12 +1,10 @@
-# external
-from dephell_links import DirLink, FileLink
-
 # app
-from .conda import CondaCloudRepo, CondaGitRepo, CondaRepo
-from .git.git import GitRepo
-from .local import LocalRepo
-from .release import ReleaseRepo
-from .warehouse import WareHouseRepo
+from ._conda import CondaCloudRepo, CondaGitRepo, CondaRepo
+from ._getter import get_repo
+from ._git.git import GitRepo
+from ._local_single import LocalRepo
+from ._release import ReleaseRepo
+from ._warehouse import WarehouseSimpleRepo, WarehouseAPIRepo, WarehouseLocalRepo
 
 
 __all__ = [
@@ -15,26 +13,9 @@ __all__ = [
     'CondaRepo',
     'get_repo',
     'GitRepo',
+    'LocalRepo',
     'ReleaseRepo',
-    'WareHouseRepo',
+    'WarehouseAPIRepo',
+    'WarehouseLocalRepo',
+    'WarehouseSimpleRepo',
 ]
-
-_repos = dict(
-    conda_cloud=CondaCloudRepo,
-    conda_git=CondaGitRepo,
-    conda=CondaRepo,
-    pypi=WareHouseRepo,
-)
-
-
-def get_repo(link=None, name: str = None):
-    if name is not None:
-        return _repos[name]()
-
-    if link is None:
-        return WareHouseRepo()
-    if getattr(link, 'vcs', '') == 'git':
-        return GitRepo(link)
-    if isinstance(link, (DirLink, FileLink)):
-        return LocalRepo(path=link.short)
-    return ReleaseRepo(link=link)