+++ dephell/repositories/_git/gitlab.py
@@ -6,7 +6,7 @@ from urllib.parse import urlencode
 import requests
 
 # app
-from ..utils import cached_property
+from ..cached_property import cached_property
 from .base import BaseRepo