+++ dephell/repositories/_warehouse/__init__.py
@@ -1,0 +1,10 @@
+from ._api import WarehouseAPIRepo
+from ._local import WarehouseLocalRepo
+from ._simple import WarehouseSimpleRepo
+
+
+__all__ = [
+    'WarehouseAPIRepo',
+    'WarehouseLocalRepo',
+    'WarehouseSimpleRepo',
+]