+++ tests/test_repositories/test_git_git.py
@@ -8,7 +8,7 @@ import pytest
 from dephell_links import VCSLink
 
 # project
-from dephell.repositories.git.git import GitRepo
+from dephell.repositories import GitRepo
 
 
 loop = asyncio.get_event_loop()